Move IDictionaryTests
